processWrite method for reuse in write call
